Minor changes.
git-svn-id: svn://svn.zabbix.com/trunk@2226 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/libs/zbxsysinfo/common/common.c
@@ -1697,7 +1697,7 @@ int	CHECK_DNS(const char *cmd, const char *ip_and_zone,double  *value, const cha
 	{
 		if(strlen(ip_and_zone)>0)
 		{
-			if(isdigit(ip_and_zone[strlen(ip_and_zone)-1]))
+			if(isdigit((int)ip_and_zone[strlen(ip_and_zone)-1]))
 			{
 				strcpy(ip,ip_and_zone);
 				strcpy(zone,"localhost");
